culture day padding and grid display
fixed the right padding on culture day images so they all align nicely.

Also inserted separate culture day posts and figured out how to display them in a grid that somewhat aligns. This knowledge can be used in the future, however, for now, I will stick with the regular images as they are because that works with that right padding