Scrapbook full thumbnail and embeded youtube video
Excluded the scrapbook thumbnail image from the bootstrap container to allow it to expand to the full width of the screen. To maintain the container structure for the rest of the page layout, wrapped the remaining content above and below the image in two separate divs with the container class.

Used <iframe> to embed the youtube video into my page and "ratio" class to ensure it sizes accordingly to the screen size and snapshot image columns while maintaining the correct video ratio.