Code cleanup before pull request
Removed commented out code from personal branding that contained copy-pasted code from a different project.

Removed extra top padding on scrapbook motion video project. But added a comment to tell where it needs to be put back in, in case the html dom layout changes in future back to how it was a few pushes back.