More material for personal branding
Added logos, pattern and mockups and social media image grid.

Created two divs for most items and hid one of the two divs either on a big screen or a small screen to respectively show them filling the full screen outside of the container or having margins inside the container respective of the screen size.