homepage recent project section
Added a recent project section on the homepage that lets the viewer easily access my favourite recent projects and the all project page.

This change also meant showing the footer links not only on the small screen, when they disappear from the main homepage-grid-div text section, but also on the bigger screen sizes. This is now needed since the content on the homepage has grown and requires a scroll to access all parts.

Updated some text on home and aboput page to reflect me starting my third year not second